update for ubuntu.

+++ bin/install_python.sh
@@ -7,7 +7,16 @@ if [ -f ~/.wgetrc ]; then
 fi
 echo "check_certificate = off" >> ~/.wgetrc
 
-sudo yum install bzip2 bzip2-libs bzip2-devel patch -y
+if [ -f /etc/redhat-release ]; then
+  sudo yum install bzip2 bzip2-libs bzip2-devel patch -y
+elif [ -f /etc/debian_version ]; then
+  sudo apt install build-essential libbz2-dev libdb-dev \
+    libreadline-dev libffi-dev libgdbm-dev liblzma-dev \
+    libncursesw5-dev libsqlite3-dev libssl-dev \
+    zlib1g-dev uuid-dev tk-dev \
+    libgnutls28-dev \
+    -y
+fi
 
 cd ~
 git clone git://github.com/yyuu/pyenv.git .pyenv
@@ -19,8 +28,8 @@ git clone git://github.com/yyuu/pyenv-virtualenv.git
 
 cd ~
 mkdir .pyenv/cache
-pyenv install 2.7.8
-pyenv global 2.7.8
+pyenv install 3.8.3
+pyenv global 3.8.3
 
 if [ -f ~/tmp/.wgetrc.backup ]; then
   mv ~/tmp/.wgetrcbackup ~/.wgetrc
